[tools] Add tool 'perf table_scan'
This is useful to show row count and scanning time cost of
tablets in a table, we can use it to check whether there is
a row count skew in different tablets, or to check whether
there is a long tail to scan different tablets.

And also we do some refactor for tool 'perf loadgen' in this commit.

Change-Id: Ibebed05d9a91ae3f126d3cf2d92f66852b253edd
Reviewed-on: http://gerrit.cloudera.org:8080/12516
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>