[tools] Add tool to copy table data to another table
Copy table data to another table; the two tables could be in the same
cluster or not. The two tables must have the same table schema, but
could have different partition schemas. Alternatively, the tool can
create the new table using the same table and partition schema as the
source table.

It's useful to use this tool to copy small tables, or just copy table
schemas in a convenient way, and it's recommend to use Java client's
Backup/Restore features to copy large tables.

Change-Id: Ifdec51701ac9ec57739b1a6f7c18786294642a54
Reviewed-on: http://gerrit.cloudera.org:8080/12563
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>