KUDU-2348: Pick a random replica in RemoteTablet.java
In RemoteTablet.java, 'getClosestServerInfo' always returns
whichever server ends up last in the map iteration order, if
all servers in the cluster is fully remote. This behavior
would cause heavy load on a particular server if there are
few servers in the cluster.

This patch fixes the issue by generating a random non-negative
integer(randomInt) in RemoteTablet initializer and choose server
based on randomInt. The randomInt variable is static so the
choice is same across multiple RemoteTablet instances.

Change-Id: I3d70e45d4c9532bb32223c1dddd0936b4ff8fd99
Reviewed-on: http://gerrit.cloudera.org:8080/12158
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>