[ksck] Filter tables and tablets in KsckCluster
The ksck tool executes slowly if there are too many tables in a cluster
even if we just want to check some specific tables and tablets.
Filtering tables and tablets in KsckCluster and only tracking specified
tables and tablets would speed up the execution of ksck tool with
'--tables=<tables>' and '--tablets=<tablets>' args.

Change-Id: I23b6e6ef258d3498a42af7f92b63392a59c99761
Reviewed-on: http://gerrit.cloudera.org:8080/13937
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>