KUDU-2934: add MeanGauge type to metrics
The current merge method is not suitable for the metric
'average_diskrowset_height', it is a mean value and we
could not just sum up the metric values when merging
metrics.

There is another issue with the metric 'average_diskrowset_height',
the value of the metric is double but it was used as an AtomicGauge
metric, which is only for types that are convertible to/from int64_t.

This patch added a new 'MeanGauge' type for metrics which show
the mean value of something.

Change-Id: I827d3acb5d4f47a7e28ac40cb7df392c29c82a40
Reviewed-on: http://gerrit.cloudera.org:8080/14252
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>