rebalancer: fix on Runner::UpdateMovesInProgressStatus()
This patch added a new parameter 'has_pending_moves' for
Runner::UpdateMovesInProgressStatus(), see the comments for
this method.

Before this patch, the return value of this method was ambiguous,
and there was a bug in PolicyFixer: When no pending operation
completed, the rebalancer tool would try to get a new set of moves
from the algorithm and perform it. Now, the rebalancer would do this
only in case of an error or no pending movement.

Change-Id: I8a25574d13f36a8f4f98f4340aa2086711b1c741
Reviewed-on: http://gerrit.cloudera.org:8080/14742
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins