[client] support resolve one master address to multiple addresses
The master addresses of kudu client are fixed during runtime, we need to
restart the client if we want to change the master addresses.

In some network settings we can map a domain name to multiple ip addresses,
if we configure the client with such a domain name as a 'master address',
resolve it and try to connect each ip address, we can easily change the master
addresses by network settings modifications.

Change-Id: Ie242bc1331596902fa16f1c6d1b439d78b73977a
Reviewed-on: http://gerrit.cloudera.org:8080/15036
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>