[java] fix bug in getClosestServerInfo code
When client handles tablet not found error, it removes the tablet server
from the RemoteTablet's locations. If all tablet servers are removed,
the code in getClosestServerInfo will throw a division by zero exception.
This patch fixed it and added a test to verify getClosestServerInfo
returns null when all locations of a tablet are invalid.

Change-Id: Ib4dc471b5044ea4b5bd6202ccd2a707d6e229ea0
Reviewed-on: http://gerrit.cloudera.org:8080/15444
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins